refactor(rocm): move the ROCm benchmark helpers out of the upstream harness
flashinfer_benchmark_utils.py had accumulated six ROCm-only definitions
that nothing on the CUDA path calls. They move to a new
benchmarks/routines/rocm_utils.py, and the upstream file keeps only the
IS_HIP branch in filter_backends_by_compute_capability plus a re-export
so attention.py's import list is unchanged:

    flashinfer_benchmark_utils.py  +141/-6  ->  +81/-6

routines/ rather than the existing rocm_benchmarks/ directory, which has
no __init__.py -- it holds standalone scripts, so importing from it would
mean turning it into a package.

Four helpers stay behind despite also being ROCm-motivated:
l2_flush_size_mb, bench_timing_kwargs, as_nhd_paged_kv_cache and
record_backend_resolution are called from moe.py and gemm.py on the CUDA
path as well.

Checked on gfx950: get_device_arch reports gfx950, rocm_supported_backends
returns ['fa2', 'auto'] for batch decode and [] for an unknown routine,
and filter_backends_by_compute_capability still strips trtllm-gen with the
architecture-worded warning.
